Browserify cleaning; preparing for lateload
