+++ src/index.html
@@ -8,7 +8,6 @@
     <meta name="viewport" content="width=device-width">
     <link rel="shortcut icon" href="/favicon.ico">
     <link rel="stylesheet" href="styles/main.css">
-    <link href='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.css' rel='stylesheet' />
   </head>
 
   <body>
@@ -16,7 +15,6 @@
       <input id="inp" placeholder="Geocode an address...">
     </form>
     <div id="map"></div> 
-    <script src='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.js'></script>
     <script src="https://rawgit.com/wpears/grassjs/master/grass.js"></script>
     <script src="scripts/bundle.js"></script>
   </body>
