+++ src/scripts/main.js
@@ -1,7 +1,12 @@
 'use strict';
 
 var loadLate = require('./loadLate');
+var subAndRun = require('./subAndRun');
 
+var subHub = subAndRun();
+
+var mapboxJS ='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.js';
+var mapboxCSS='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.css';
 L.mapbox.accessToken = 'pk.eyJ1IjoiY2ZwYiIsImEiOiJodmtiSk5zIn0.VkCynzmVYcLBxbyHzlvaQw';
 var map = L.mapbox.map('map', 'mapbox.streets').setView([38, -122], 10);
 var features = L.mapbox.featureLayer(null).addTo(map);
@@ -24,42 +29,19 @@ function flipCoords(arr){
   return [arr[1],arr[0]];
 }
 
-var handler = function(){
-  var subscribers = [];
-  
-  function subscribe(fn){
-    return subscribers.push(fn);
-  } 
-
-  function unsubscribe(fn){
-    for(var i=0; i<subscribers.length; i++){
-      if(subscribers[i] === fn){
-        return subscribers.splice(i,1);
-      }
-    }
+function process(err, result){
+  if(err){
+    result = randomBayPoint(); 
   }
+  var json = JSON.parse(result);
+  subHub.run(json);
+}
 
-  function process(err, result){
-    if(err){
-      result = randomBayPoint(); 
-    }
-    var json = JSON.parse(result);
-    for(var i=0; i<subscribers.length; i++){
-      subscribers[i](json);
-    }
-  }
-   
-  return {
-    process:process,
-    subscribe:subscribe,
-    unsubscribe:unsubscribe
-  };
-}();
 
-handler.subscribe(function(result){
+subHub.subscribe(function(result){
   console.log(result);
 })
-handler.subscribe(function(result){
+subHub.subscribe(function(result){
   features.setGeoJSON(result);
   map.panTo(flipCoords(result.coordinates));
 });
@@ -69,7 +51,7 @@ inp.addEventListener('keydown',function(e){
   if(e.keyCode === 13){
     var query = this.value;
     this.value = '';
-    grass.geocode(query, handler.process);
+    grass.geocode(query, process);
   }
 });
